Remove FAQ about companies printing decks
Removed because, not sure the text is consistent with actual license, or OWASP brand guidance. Originally this text was added to help explain why Blackfoot could print copies of the deck and distribute it free of charge. Other commercial orgs are printing and selling decks, and probably don't comply with the text of this former FAQ.